% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{package load and attach events}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{internal}

\item{pkgname}{internal}
}
\description{
package load and attach events
}
