% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilioapi.R
\name{getTwilioMessagingNums}
\alias{getTwilioMessagingNums}
\alias{getTwilioMessages}
\alias{sendTwilioMessagingService}
\alias{buyTwilioNumber}
\alias{setupTwilioNumber}
\alias{cycleNums}
\title{TWILIO + POLARIS}
\usage{
getTwilioMessagingNums(ms, sid, token)

getTwilioMessages(date = NULL, from = NULL, to = NULL, sid, token)

sendTwilioMessagingService(to, message, mediaUrl = NA, ms, sid, token)

buyTwilioNumber(sid, apikey, area_code)

setupTwilioNumber(sid, apikey, phone_sid)

cycleNums(sid, apikey, area_code, num)
}
\arguments{
\item{ms}{TBD}

\item{sid}{TBD}

\item{token}{TBD}

\item{date}{TBD}

\item{from}{TBD}

\item{to}{TBD}

\item{message}{TBD}

\item{mediaUrl}{TBD}

\item{apikey}{TBD}

\item{area_code}{TBD}

\item{phone_sid}{TBD}

\item{num}{TBD}
}
\description{
TBD
}
\section{Functions}{
\itemize{
\item \code{getTwilioMessagingNums}: Get the phone numbers associated with a messaging service

\item \code{getTwilioMessages}: Get all text message for a certain criteria

\item \code{sendTwilioMessagingService}: Send a text message via a messaging service

\item \code{buyTwilioNumber}: buys a single number

\item \code{setupTwilioNumber}: sets up a single phone number

\item \code{cycleNums}: cycle nums
}}

